fastapi_backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
anthropic
pydantic
aiohttp
python-dotenv
groq